phy/lpddr4: extract io helpers to a separate class
